[20] fixup sql queries, start on auth, omnierr
