+++ src/omni_error.rs
@@ -1,9 +1,11 @@
 #[derive(thiserror::Error, Debug)]
 pub enum OmniError {
+    #[error("AuthError: {0}")]
+    AuthError(#[from] crate::users::auth::AuthError),
+    #[error("PhotoUrlError: {0}")]
+    PhotoUrlError(#[from] crate::users::photourl::PhotoUrlError),
     #[error("sqlx::Error: {0}")]
     SqlxError(#[from] sqlx::Error),
-    #[error("users::photourl::PhotoUrlError: {0}")]
-    PhotoUrlError(#[from] crate::users::photourl::PhotoUrlError),
     #[error("serde_json::Error: {0}")]
     SerdeJsonError(#[from] serde_json::Error),
 }
